requirements/common.txt
Django==3.2.14
djangorestframework
Pillow
django-cors-headers

aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
backoff==1.11.1
certifi==2022.6.15
charset-normalizer==2.0.12
Deprecated==1.2.13
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
httplib2==0.20.4
idna==3.3
oauthlib==3.2.0
packaging==21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
redis==4.3.3
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
spotipy==2.20.0
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
yarl==1.7.2
djangorestframework-simplejwt
django-allauth
dj-rest-auth